+++ src/main/java/io/aeolabs/sonar/ts/PrometheusHandler.java
@@ -1,0 +1,20 @@
+package io.aeolabs.sonar.ts;
+
+
+public class PrometheusHandler implements TimeSerial {
+
+
+
+
+    @Override
+    public void addMetric(TsMetric tsMetric) {
+//        this.gauges.get(measure.getMetric())
+//                .labels(project.getKey(), project.getName())
+//                .set(Double.parseDouble(measure.getValue()));
+    }
+
+    @Override
+    public void updateGauges() {
+
+    }
+}