+++ src/main/java/io/aeolabs/sonar/ts/TimeSerial.java
@@ -1,0 +1,8 @@
+package io.aeolabs.sonar.ts;
+
+
+public interface TimeSerial {
+    public void addMetric(TsMetric tsMetric);
+    public void updateGauges();
+
+}