+++ src/main/java/io/aeolabs/sonar/ts/TsMetric.java
@@ -1,0 +1,13 @@
+package io.aeolabs.sonar.ts;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class TsMetric {
+    private final String metric;
+    private final String key;
+    private final String projectName;
+    private final Double value;
+}